dashboard(filtering): move mempool metrics to Overview, compact What's active, reorder
