+++ dashboard/src/app/(dashboard)/mempool/page.tsx
@@ -8,10 +8,8 @@ import { Badge } from "@/components/ui/Badge";
 import { SkeletonCard } from "@/components/ui/Skeleton";
 import { StatCard } from "@/components/ui/StatCard";
 import { SectionErrorBoundary } from "@/components/ui/SectionErrorBoundary";
-import { TimeSeriesChart } from "@/components/ui/MiniChart";
 import { useConfig } from "@/hooks/queries/useConfigQueries";
 import { useReaperStatus } from "@/hooks/queries";
-import { useMempoolSeries } from "@/hooks/useMempoolSeries";
 import { type ReaperStats } from "@/lib/api/reaper";
 import { fetchApi } from "@/lib/api/client";
 
@@ -153,12 +151,6 @@ export default function MempoolPage() {
         actions={<Badge variant="success">rpc · lightweight</Badge>}
       />
 
-      {/* Tx-count tiles + the mempool-over-time graph. Sits above the embed so
-          the two headline counts and the trend are the first thing you see. */}
-      <SectionErrorBoundary section="Mempool metrics">
-        <MempoolMetrics mempool={mempool} />
-      </SectionErrorBoundary>
-
       {/* The real mempool.space UI of THIS node's own mempool, served
           same-origin through the dashboard, at the very top. The lightweight
           RPC cards and the Reaper filter breakdown follow below. */}
@@ -186,72 +178,6 @@ export default function MempoolPage() {
           </SectionErrorBoundary>
         </>
       )}
-    </div>
-  );
-}
-
-// ─── mempool tx-count tiles + over-time graph ────────────────────────────────
-
-/**
- * Two headline tiles (current mempool tx count, current block-template tx count)
- * plus a "mempool over time" line chart.
- *
- * Data comes from {@link useMempoolSeries}: the node's server-side `/pool/series`
- * ring (30s samples, ~24h, survives reloads) when it carries the mempool fields,
- * otherwise a live in-browser session buffer keyed on this page's own
- * `/buds/mempool` poll. The "Mempool" tile prefers the live `total` already
- * fetched on the page and only falls back to the latest series sample; the
- * "This block" tile is server-ring only (the RPC view has no block-template
- * count) and shows "—" until the ring has it.
- */
-function MempoolMetrics({ mempool }: { mempool?: BudsMempool }) {
-  const series = useMempoolSeries();
-
-  // Live getmempoolinfo.size when we have it, else the newest series sample.
-  const mempoolNow = mempool?.total ?? series.latestMempoolTxs ?? undefined;
-  const blockNow = series.latestBlockTxs ?? undefined;
-
-  return (
-    <div className="space-y-4">
-      <div className="grid grid-cols-2 md:grid-cols-4 gap-4">
-        <StatCard
-          label="Mempool"
-          value={formatCount(mempoolNow)}
-          sublabel="transactions right now"
-          tooltip="getmempoolinfo.size — transactions your node currently holds"
-        />
-        <StatCard
-          label="This block"
-          value={formatCount(blockNow)}
-          sublabel="in the block template"
-          tooltip="transactions in the block template your node is currently building"
-        />
-      </div>
-
-      <Card>
-        <div className="mb-3 flex items-start justify-between gap-2">
-          <div>
-            <h3 className="t-title" style={{ color: "var(--fg)" }}>
-              Mempool over time
-            </h3>
-            <p className="t-caption" style={{ color: "var(--dim)", marginTop: "2px" }}>
-              Transactions in your node&apos;s mempool, sampled over time
-            </p>
-          </div>
-          <Badge variant="default">{series.serverBacked ? "history" : "live (session)"}</Badge>
-        </div>
-        <TimeSeriesChart
-          data={series.mempoolTxs}
-          minZero
-          ariaLabel="Mempool transaction count over time"
-          formatValue={(v) => Math.round(v).toLocaleString()}
-        />
-        <p className="t-caption" style={{ color: "var(--fainter)", marginTop: "12px" }}>
-          {series.serverBacked
-            ? "Drawn from this node's server-side history (sampled every 30s, up to 24h) so it survives reloads."
-            : `Collecting live in-browser samples (${series.sampleCount} this session) — the node's server-side history will back this chart once it has accumulated a few. Resets on reload.`}
-        </p>
-      </Card>
     </div>
   );
 }
